+++ example-notes-app.html
@@ -5,9 +5,9 @@
     <title></title>
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css"
           integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
-    <script src="lib/serviceworker/register.js"></script>
+    <script src="src/serviceworker/register.js"></script>
     <script type="module">
-      import App from './lib/app.js'
+      import App from './src/app.js'
 
       new App('examples/notes-app', { api: 'http://localhost:3000' })
     </script>
