--- lib/serviceworker/register.js
@@ -1,5 +1,0 @@
-if ('serviceWorker' in navigator) {
-  window.addEventListener('load', () => {
-    navigator.serviceWorker.register('/lib/serviceworker/worker.js').then(() => {}).catch(() => {});
-  });
-}
