+++ src/serviceworker/register.js
@@ -1,0 +1,5 @@
+if ('serviceWorker' in navigator) {
+  window.addEventListener('load', () => {
+    navigator.serviceWorker.register('/src/serviceworker/worker.js').then(() => {}).catch(() => {});
+  });
+}
